evaluation-service/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# ragas==0.1.5
# langchain-ollama==0.1.0
# langchain-core==0.2.40
# python-multipart==0.0.6
# pydantic==2.5.0
# python-dotenv==1.0.0
# pandas==2.1.4
# numpy==1.24.3
# requests==2.31.0
# datasets==2.14.0
# pyarrow

fastapi==0.104.1
uvicorn[standard]==0.24.0
ragas==0.1.5
langchain-ollama==0.1.0      # keep if you may want OllamaEmbeddings
langchain-core==0.2.40
langchain-community==0.2.10  # ✅ add this (needed for Ollama LLM class)
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
datasets==2.14.0
pyarrow
